amazon-project/requirements.txt
python-dotenv==1.0.1
requests==2.32.3
streamlit==1.37.0
loguru==0.7.2
